feat: auto-capture OAuth tokens in gateway proxy, eliminate secret leaks
Route all Salesforce HTTP traffic through the gateway proxy with secret
placeholder resolution. The gateway auto-detects OAuth2 token responses
(by checking for access_token in the JSON body), stores captured fields
in the encrypted secret store, and strips the entire response body —
tokens never reach the script or LLM context.

Key changes:
- Extract HTTP proxy into gateway-http-proxy.ts with SSRF guard, secret
  placeholder resolution, bearer injection, and OAuth auto-capture
- captureResponseFields API lets callers declare which response fields
  to store as secrets (defaults to access_token + refresh_token)
- Remove store_secret/subprocess/secret-command from salesforce script
- Script uses <secret:SF_INSTANCE_URL> placeholder for all API calls
- Fix delegation process leak: cleanup on completion, reuse session ID
